+++ src/commands.cpp
@@ -318,10 +318,12 @@ void Commands::testPari() {
     if (!formula.first) {
       continue;
     }
-    Log::get().info(seq.id_str() + ": " + formula.second.toString(true));
+    auto pariCode = formula.second.toString(true);
+    Log::get().info(seq.id_str() + ": " + pariCode);
     Sequence expSeq;
     size_t numTerms = seq.existingNumTerms();
-    if (inceval.init(program)) {
+    if (inceval.init(program) ||
+        pariCode.find("binomial") != std::string::npos) {
       numTerms = std::min<size_t>(numTerms, 10);
     }
     evaluator.eval(program, expSeq, numTerms);